Refactor AI and API key services
