+++ backend/src/agents/news-analyst.ts
@@ -1,7 +1,7 @@
 import type { FastifyBaseLogger } from 'fastify';
 import { getNewsByToken } from '../repos/news.js';
 import { insertReviewRawLog } from '../repos/review-raw-log.js';
-import { callAi, extractJson } from '../util/ai.js';
+import { callAi, extractJson } from '../services/ai.js';
 import { isStablecoin } from '../util/tokens.js';
 import type { RebalancePrompt, RunParams } from './main-trader.types.js';
 import {
