+++ backend/src/routes/exchange-api-keys.ts
@@ -6,11 +6,8 @@ import {
   setBinanceKey,
   clearBinanceKey,
 } from '../repos/exchange-api-keys.js';
-import {
-  ApiKeyType,
-  verifyApiKey,
-  encryptKey,
-} from '../util/api-keys.js';
+import { verifyBinanceKey } from '../services/binance.js';
+import { encryptKey } from '../util/crypto.js';
 import { errorResponse, ERROR_MESSAGES } from '../util/errorMessages.js';
 import { REDACTED_KEY } from './_shared/constants.js';
 import { getValidatedUserId, userPreHandlers } from './_shared/guards.js';
@@ -57,7 +54,7 @@ async function verifyAndSave(
     reply: FastifyReply,
     userId: string
 ) {
-  const verRes = await verifyApiKey(ApiKeyType.Binance, key, secret);
+  const verRes = await verifyBinanceKey(key, secret);
   if (verRes !== true) return reply.code(400).send(errorResponse(formatVerificationError(verRes)));
   const encKey = encryptKey(key);
   const encSecret = encryptKey(secret);
